+++ CHANGELOG.md
@@ -10,6 +10,16 @@ Note that there may be gaps in the version history, which may happen if a releas
 
 # Current Version
 
+## v5.2.3 (August 4, 2025)
+
+- fixed bug that prevented renaming a file by adding text to the end of the filename.
+
+# Previous Versions
+
+## v5.2.2 (August 3, 2025)
+
+- fixed bug that would show dot-dirs as both a directory and a (broken) file entry.
+
 ## v5.2.0 (July 15, 2025)
 
 Granting events now returns the element(s) involved:
@@ -21,8 +31,6 @@ Granting events now returns the element(s) involved:
 - granting a `delete` event returns an _array_ of entries, containing either a single element at index 0 if a file got deleted, or the dir entry at index 0 _plus all descendants_ in the subsequent indices, if a dir got deleted.
 
 Additionally, the `show-top-level` attribute was missing CSS to change the styling, which has been fixed. This introduces the `--highlight-background-bw` CSS variable for controlling the color of the button bar when the top level "." path is hidden.
-
-# Previous Versions
 
 ## v5.0.0 (August 11, 2024)